export start

+++ include/minishell.h
@@ -6,7 +6,7 @@
 /*   By: akharraz <akharraz@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/02 19:55:30 by tel-mouh          #+#    #+#             */
-/*   Updated: 2022/10/23 04:38:23 by akharraz         ###   ########.fr       */
+/*   Updated: 2022/10/24 12:15:49 by akharraz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -57,7 +57,7 @@ char    *lexer(char *buff, t_sub *sub);
 int		accepted(t_node *new);
 // ------------------dir------------------------
 char	*getbasename(char *path);
-size_t	getdir(char **base_name);
+int 	getdir(char **base_name);
 char	*dub_add(char *str, char *ad);
 // -----------------init------------------------
 int		init(t_vars *vars, char **env);
